requirements.txt
python-dotenv
requests
google-generativeai
backoff
